services/lk-producer/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
kafka-python==2.0.2
prometheus-flask-exporter==0.22.4
requests==2.31.0